teamskills/backend/requirements.txt
# python==3.12.7
python_dotenv==0.21.0
pdfplumber==0.11.7
google-cloud-vision==3.10.2
pdf2image==1.17.0
pillow==11.3.0